+++ src/test/java/org/dataone/bookkeeper/jdbi/QuotaDAOTest.java
@@ -50,7 +50,7 @@ public class QuotaDAOTest extends BaseTestCase {
     @DisplayName("Test get quota by quota ID")
     public void testGetQuotaById() {
 
-        assertTrue(quotaDAO.findQuotaById(1L).size() == 1);
+        assertTrue(quotaDAO.findQuotaById(1).size() == 1);
     }
 
     /**
@@ -70,31 +70,31 @@ public class QuotaDAOTest extends BaseTestCase {
             "VALUES " +
             "(?, ?, ?, ?, ?, ?)", "quota", "test_customer_quota", 54321, 543210, "byte", 500);
         });
-        assertTrue(quotaDAO.findQuotasByCustomerId(500L).size() == 2);
-        assertThat(quotaDAO.findQuotasByCustomerId(0L).isEmpty());
+        assertTrue(quotaDAO.findQuotasByCustomerId(500).size() == 2);
+        assertThat(quotaDAO.findQuotasByCustomerId(0).isEmpty());
     }
 
     @Test
     @DisplayName("Test inserting a quota")
     public void testInsert() {
-        quotaDAO.insert("quota", "test_quota", 12345L,
-            123450L, "byte", 500L);
-        assertThat(quotaDAO.findQuotasByCustomerId(500L).size() > 0);
+        quotaDAO.insert("quota", "test_quota", 12345,
+            123450, "byte", 500);
+        assertThat(quotaDAO.findQuotasByCustomerId(500).size() > 0);
 
     }
 
     @Test
     @DisplayName("Test updating a quota")
     public void testUpdate() {
-        quotaDAO.insert("quota", "test_quota", 12345L,
-            123450L, "byte", 400L);
-        Quota quota = quotaDAO.findQuotasByCustomerId(400L).get(0);
-        Long quotaId = quota.getId();
-        quotaDAO.update(quotaId, "quota", "test_quota_2", 56789L,
-            567890L, "byte", 400L);
-        assertThat(quotaDAO.findQuotasByCustomerId(400L).get(0).getName() == "test_quota_2");
-        assertThat(quotaDAO.findQuotasByCustomerId(400L).get(0).getSoftLimit() == 56789);
-        assertThat(quotaDAO.findQuotasByCustomerId(400L).get(0).getHardLimit() == 567890);
+        quotaDAO.insert("quota", "test_quota", 12345,
+            123450, "byte", 400);
+        Quota quota = quotaDAO.findQuotasByCustomerId(400).get(0);
+        Integer quotaId = quota.getId();
+        quotaDAO.update(quotaId, "quota", "test_quota_2", 56789,
+            567890, "byte", 400);
+        assertThat(quotaDAO.findQuotasByCustomerId(400).get(0).getName() == "test_quota_2");
+        assertThat(quotaDAO.findQuotasByCustomerId(400).get(0).getSoftLimit() == 56789);
+        assertThat(quotaDAO.findQuotasByCustomerId(400).get(0).getHardLimit() == 567890);
 
     }
 }
